refactor: 페이지 단위 콜로케이션 구조 적용
- 기존 src/features/*, src/hooks/* 파일을 app 하위로 이동
- 불필요한 .gitkeep 정리
