+++ src/app/label/[imageKey]/_components/label-step/label-step2.tsx
@@ -2,10 +2,10 @@
 
 import { useMemo } from 'react';
 
+import useInteractionTimer from '@/app/label/[imageKey]/_hooks/use-interaction-timer';
+import useUpdateLabel from '@/app/label/[imageKey]/_hooks/use-update-label';
 import BottomCTA from '@/components/ui/bottom-cta';
 import ButtonList from '@/components/ui/button-list';
-import useInteractionTimer from '@/hooks/use-interaction-timer';
-import useUpdateLabel from '@/hooks/use-update-label';
 import { TOTAL_LABELING_STEPS } from '@/lib/constants';
 import {
   GetLabelStatusResponse,
