+++ src/app/label/[imageKey]/_hooks/use-interaction-timer.ts
